requirements.txt
lumibot
ta
sklearn
google-cloud-secret-manager==1.0.0